#7- BOJ 1152 문제 세팅
