+++ Algorithm/Algorithm/BOJ/main.swift
@@ -1,17 +1,9 @@
 //
-//  2588.swift
-//  Algorithm - 곱셈
+//  1152.swift
+//  Algorithm
 //
-//  Created by 김인영 on 2023/09/09.
+//  Created by 김인영 on 2023/09/10.
 //
 
 import Foundation
 
-var A = Int(readLine()!)!
-var B = Int(readLine()!)!
-var result = 0
-
-print(A * ( B % 10))
-print(A * (B % 100 / 10))
-print(A * (B / 100))
-print(A * B)
